fix imports and exports - cleanup and cosmetic
