+++ src/Clients/nodejs/package.json
@@ -1,7 +1,8 @@
 {
     "name": "@uipath/coreipc",
     "version": "1.0.0",
-    "main": "src/index.ts",
+    "main": "dist/index.js",
+    "typings": "dist/index",
     "scripts": {
         "refreshVSToken": "vsts-npm-auth -config .npmrc",
         "build": "tsc",
@@ -60,6 +61,7 @@
         "concurrently": "4.1.0",
         "fs-extra": "9.0.1",
         "istanbul-instrumenter-loader": "3.0.1",
+        "madge": "^3.9.2",
         "mocha": "6.1.4",
         "mocha-junit-reporter": "1.23.1",
         "mocha-multi-reporters": "^1.1.7",
@@ -79,7 +81,7 @@
         "tslint-eslint-rules": "5.4.0",
         "typedoc": "^0.15.0",
         "typedoc-plugin-markdown": "1.2.0",
-        "typescript": "3.9.5",
+        "typescript": "3.9.7",
         "uuid": "8.0.0"
     }
 }
