+++ src/Clients/nodejs/src/core/ipc/annotations/ClassAnnotationsWrapper.ts
@@ -1,5 +1,5 @@
 // tslint:disable: no-namespace
-import { PublicCtor } from '@foundation';
+import { PublicCtor } from '../../../foundation';
 import { IIpc } from '..';
 import { Ipc } from '../Ipc';
