+++ src/Clients/nodejs/src/core/ipc/annotations/MethodAnnotationsWrapper.ts
@@ -1,4 +1,4 @@
-import { PublicCtor, Primitive, InvalidOperationError } from '@foundation';
+import { PublicCtor, Primitive, InvalidOperationError } from '../../../foundation';
 import { IIpc } from '..';
 import { Ipc } from '../Ipc';
