+++ src/Clients/nodejs/src/core/ipc/callback-store/CallbackStore.ts
@@ -1,5 +1,5 @@
 import { IIpc } from '../IIpc';
-import { PublicCtor, argumentIs } from '@foundation';
+import { PublicCtor, argumentIs } from '../../../foundation';
 
 /* @internal */
 export class CallbackStore implements IIpc.CallbackStoreInternal {
