+++ src/Clients/nodejs/src/core/ipc/config-store/ConfigNode.ts
@@ -1,4 +1,4 @@
-import { TimeSpan, Timeout, ConnectHelper } from '@foundation';
+import { TimeSpan, Timeout, ConnectHelper } from '../../../foundation';
 
 /* @internal */
 export interface ConfigNode {
