+++ src/Clients/nodejs/src/core/ipc/config-store/ConfigStore.ts
@@ -6,7 +6,7 @@ import {
     TimeSpan,
     ArgumentError,
     ConnectHelper,
-} from '@foundation';
+} from '../../../foundation';
 
 import {
     ConfigAction,
