+++ src/Clients/nodejs/src/core/ipc/contract-store/ContractStore.ts
@@ -1,4 +1,4 @@
-import { PublicCtor } from '@foundation';
+import { PublicCtor } from '../../../foundation';
 import { IIpc } from '../IIpc';
 import { ServiceInfo } from '.';
