+++ src/Clients/nodejs/src/core/ipc/contract-store/OperationInfo.ts
@@ -1,5 +1,5 @@
 import 'reflect-metadata';
-import { PublicCtor, Primitive, CancellationToken } from '@foundation';
+import { PublicCtor, Primitive, CancellationToken } from '../../../foundation';
 import { IIpc } from '../IIpc';
 
 /* @internal */
