+++ src/Clients/nodejs/src/core/ipc/dtos/Message.ts
@@ -1,24 +1,22 @@
 // tslint:disable: variable-name
 
-import { TimeSpan } from '@foundation';
+import { TimeSpan } from '../../../foundation';
 
 export class Message<T = void> {
-    constructor(
-        public readonly payload?: T,
+    constructor(args?: {
+        payload?: T,
         requestTimeout?: TimeSpan,
-    ) {
-        this.requestTimeout = requestTimeout ?? TimeSpan.zero;
+    }) {
+        this.payload = args?.payload;
+        this.requestTimeout = args?.requestTimeout ?? TimeSpan.zero;
     }
 
-    public readonly requestTimeout: TimeSpan;
+    public readonly payload?: T;
+    public requestTimeout: TimeSpan;
 
     public toJSON(): unknown {
-        if (this.payload === undefined) {
-            return {};
-        } else {
-            return {
-                Payload: this.payload,
-            };
-        }
+        return this.payload !== undefined
+            ? { Payload: this.payload }
+            : {};
     }
 }
