+++ src/Clients/nodejs/src/core/types/ConnectionHookDelegate.ts
@@ -1,4 +1,4 @@
 import { IConnection, ConnectionFactoryDelegate } from '.';
-import { CancellationToken } from '@foundation';
+import { CancellationToken } from '../../foundation';
 
 export type ConnectionHookDelegate = (defaultFactory: ConnectionFactoryDelegate, ct: CancellationToken) => Promise<IConnection | void>;
