+++ src/Clients/nodejs/src/core/types/MethodCallHookDelegate.ts
@@ -1,3 +1,3 @@
-import { CancellationToken } from '@foundation';
+import { CancellationToken } from '../../foundation';
 
 export type MethodCallHookDelegate = (methodName: string, newConnection: boolean, ct: CancellationToken) => Promise<void>;
