+++ src/Clients/nodejs/src/foundation/named-pipes/ConnectContext.ts
@@ -1,0 +1,8 @@
+import { CancellationToken, TimeSpan } from '..';
+
+export interface ConnectContext {
+    readonly pipeName: string;
+    readonly timeout: TimeSpan;
+    readonly ct: CancellationToken;
+    readonly tryConnect: () => Promise<boolean>;
+}
