+++ src/Clients/nodejs/src/foundation/named-pipes/index.ts
@@ -4,3 +4,4 @@ export * from './SocketStream';
 export * from './SocketLike';
 export * from './NamedPipeClientSocket';
 export * from './ConnectHelper';
+export * from './ConnectContext';
