+++ src/Clients/nodejs/src/foundation/threading/promises/PromiseCompletionSource.ts
@@ -1,4 +1,4 @@
-import { argumentIs } from '@foundation';
+import { argumentIs } from '../../../foundation';
 import {
     IPromiseCompletionSourceInternal,
     PromiseCompletionSourceInternal,
