+++ src/Clients/nodejs/src/foundation/threading/synchronization/AutoResetEvent.ts
@@ -3,7 +3,7 @@ import {
     PromiseCompletionSource,
     CancellationToken,
     CancellationTokenRegistration,
-} from '@foundation';
+} from '../../../foundation';
 
 /* @internal */
 export class AutoResetEvent {
