+++ src/Clients/nodejs/src/index.ts
@@ -1,3 +1,21 @@
-export * from '@core';
-export * from '@foundation-errors';
-export * from '@foundation-threading';
+export * from './core';
+export * from './foundation/errors';
+export * from './foundation/threading';
+
+import {
+    Trace,
+    IDisposable,
+    IAsyncDisposable,
+    TimeSpan,
+    ConnectHelper,
+    ConnectContext,
+} from './foundation';
+
+export {
+    Trace,
+    IDisposable,
+    IAsyncDisposable,
+    TimeSpan,
+    ConnectHelper,
+    ConnectContext,
+};
