+++ src/Clients/nodejs/test/unit/surface/foundation/PromiseCompletionSource.spec.ts
@@ -1,4 +1,5 @@
-import { expect, spy, constructing, calling, forInstance } from '@test-helpers';
+import { expect, spy, constructing, forInstance } from '@test-helpers';
+
 import {
     PromiseCompletionSource,
     FinalState,
